* develop: (377 commits)
  Little formatting changes for a better merging I hope.
  Some code refactoring.
  Lines and Arrows can be created by selecting startpoint and endpoint
  Added D3 library and removed link to online version of D3.
  Quick and dirty fix to make WebArena work in new versions of socket.io. Proper namespaces need to be put into place.
  small usability updates for the exit object
  small bugfixes for the object creation
  change the creation of objects: Instead of showing them in the top left corner the user can set a self-chosen position by clicking on the canvas
  Further small fixes for the directed links
  Some changes. Bug fixing and clean up.
  A lot of changes, most of them implements new objects for the mobile version.
  some fixes for the directed links
  Add more properties to the directed links + minor fixes
  Cleanup room code
  Fixed an error, where object handling crashed when objects had connections to nonexisting objects
  Fixed config issues in WebServer
  all the old behaviour of invisible objects to the new directed links + other small fixes
  Minor fix
  add directed links
  Added comments to the config file
  ...

Conflicts:
	source/Client/ObjectManager.js
	source/Client/guis.common/javascript/1.toolbar.js

+++ source/Client/guis.common/javascript/0.GUI.js
@@ -9,7 +9,7 @@ var GUI={};
  * language of the client (used by translate function)
  * @default de
  */
-GUI.currentLanguage='de';
+GUI.currentLanguage=Modules.Config.language;
 
 
 //This is called then forward or backwards-buttons are used in the browser.
@@ -47,6 +47,23 @@ GUI.translate=function(text){
 GUI.isTouchDevice = false;
 
 
+/**
+ * Variable indicates the type of gui which is currently used.
+ * Default value is 'desktop'.
+ */
+GUI.guiType = 'desktop';
+
+/**
+ * Sets the gui type on startup.
+ */
+$(function() {
+    var userAgent = navigator.userAgent;
+    if (userAgent.indexOf('iPhone') > 0 ||
+		(userAgent.indexOf('Android') > 0 && userAgent.indexOf('Mobile') > 0)) {
+		GUI.guiType = 'mobilephone';
+    }
+});
+
 
 /**
  * @deprecated still needed?
@@ -74,6 +91,9 @@ GUI.initResizeHandler = function() {
  */
 GUI.adjustContent = function(webarenaObject) {
 	
+	if (GUI.guiType != 'desktop') {
+	  return;
+	}
 	if (webarenaObject != undefined) {
 
 		if (!webarenaObject.isGraphical) return;
@@ -153,10 +173,9 @@ GUI.setRoomWidth = function(width) {
 	if (width < $(window).width()) {
 		width = $(window).width();
 	}
-	
-	$("#content").css("width", width);
-	$("#content > svg").css("width", width);
-	
+
+	$("#content").css("width", width );
+	$("#content > svg").css("width", width );
 }
 
 /**
@@ -176,7 +195,6 @@ GUI.setRoomHeight = function(height) {
 	
 	$("#content").css("height", height);
 	$("#content > svg").css("height", height);
-	
 }
 
 
@@ -317,29 +335,60 @@ GUI.initObjectDeletionByKeyboard = function() {
 			if (event.which == 8 || event.which == 46) {
 
 				event.preventDefault();
+				
+				//the cursor shows an object --> change the cursor to the default one	
+				var cursor = $("body").css('cursor');
+				if(cursor != "auto"){ 
+						
+					$("body").css( 'cursor', 'auto' );
+				}
+				else{
+					var result = confirm(GUI.translate('Do you really want to delete the selected objects?'));
 
-				var result = confirm(GUI.translate('Do you really want to delete the selected objects?'));
+					if (result) {
+						/* delete selected objects */
+						$.each(ObjectManager.getSelected(), function(key, object) {
 
-				if (result) {
-					/* delete selected objects */
-					$.each(ObjectManager.getSelected(), function(key, object) {
+							if ($(object.getRepresentation()).data("jActionsheet")) {
+								$(object.getRepresentation()).data("jActionsheet").remove();
+							}
 
-						if ($(object.getRepresentation()).data("jActionsheet")) {
-							$(object.getRepresentation()).data("jActionsheet").remove();
-						}
+							object.deleteIt();
 
-						object.deleteIt();
-
-					});
+						});
+					}
 				}
 			}
-			
 		}
 		
 	});
 	
 }
 
+
+/**
+ * add event handler for removing the cursor which represents an object by pressing the Escape-key
+ */
+GUI.initCursorDeletionByKeyboard = function() {
+	
+	$(document).bind("keydown", function(event) {
+		
+		if ($("input:focus,textarea:focus").get(0) == undefined) {
+		
+			if (event.which == 27) {
+
+				event.preventDefault();
+				
+				//remove objects from cursor 
+				var cursor = $("body").css('cursor');
+				if(cursor != "auto"){ 
+						
+					$("body").css( 'cursor', 'auto' );
+				}
+			}
+		}
+	});
+}
 
 
 /**
@@ -379,7 +428,6 @@ GUI.initObjectCopyCutPasteHandlingByKeyboard = function() {
  * add event handler for object selection (based on clicked position and layers)
  */
 GUI.initMouseHandler = function() {
-
 
 	if (GUI.isTouchDevice) {
 		
@@ -421,6 +469,91 @@ GUI.initMouseHandler = function() {
 			var contentPosition = $("#content").offset();
 			
 			var temp=event.target;
+				
+			//object creation via object symbols
+			var cursor = $("body").css('cursor');	
+			if(cursor != "auto"){
+				
+				var t = cursor.split("/");
+				var s = t[4].split(")");
+				var r = s[0];
+				
+				//special case for Firefox
+				if(r.charCodeAt(r.length-1)==34){
+					r = r.slice(0,r.length-1);
+				}
+				
+				var proto = ObjectManager.getPrototype(r);
+			
+				GUI.startNoAnimationTimer();
+						
+				proto.create({
+						"x" : event.pageX-30, 
+						"y" : event.pageY-65
+				});
+			
+				$("body").css( 'cursor', 'auto' );			
+			}
+			
+			//after selecting the arrow startpoint change the cursor text to arrow endpoint
+			if(GUI.getCursorText()==GUI.translate('Choose Arrow-Startpoint')){
+				GUI.setCursorText(GUI.translate("Choose Arrow-Endpoint"));
+							
+				var position = $('#besideMouse').position();
+				
+				$('#besideMouse').attr('title', position.left+','+position.top);
+				
+			}
+			
+			//after selecting the line startpoint change the cursor text to line endpoint
+			if(GUI.getCursorText()==GUI.translate('Choose Line-Startpoint')){
+				GUI.setCursorText(GUI.translate("Choose Line-Endpoint"));
+
+				var position = $('#besideMouse').position();
+				
+				$('#besideMouse').attr('title', position.left+','+position.top);	
+			}
+			
+			//after selecting the arrow endpoint create the arrow and set the position with GUI.setFinalPosition
+			if(GUI.getCursorText()==GUI.translate('Choose Arrow-Endpoint')){
+				GUI.setCursorText("");
+
+				var proto = ObjectManager.getPrototype('Arrow');
+			
+				GUI.startNoAnimationTimer();
+				
+				var title = $('#besideMouse').attr('title');
+				
+				var position = $('#besideMouse').position();
+				
+				$('#besideMouse').attr('title', title+','+position.left+','+position.top);
+										
+				var attributes;
+				var content;			
+					
+				ObjectManager.createObject(proto.type,attributes,content,GUI.setFinalPosition);	
+			}
+			
+			//after selecting the line endpoint create the line and set the position with GUI.setFinalPosition
+			if(GUI.getCursorText()==GUI.translate('Choose Line-Endpoint')){
+				GUI.setCursorText("");
+
+				var proto = ObjectManager.getPrototype('Line');
+			
+				GUI.startNoAnimationTimer();
+				
+				var title = $('#besideMouse').attr('title');
+				
+				var position = $('#besideMouse').position();
+				
+				$('#besideMouse').attr('title', title+','+position.left+','+position.top);
+										
+				var attributes;
+				var content;			
+					
+				ObjectManager.createObject(proto.type,attributes,content,GUI.setFinalPosition);
+			}
+			
 			
 			while (temp && !temp.dataObject) {
 				temp=$(temp).parent()[0];
@@ -429,7 +562,19 @@ GUI.initMouseHandler = function() {
 			var clickedObject=(temp)?temp.dataObject:false;
 			
 			//TODO check if this can be done similarly for touch devices
-		
+
+			if (GUI.couplingModeActive) {
+				if (event.pageX > $('#couplingBar').attr('x1') && $('#couplingBar:hover').length == 0) {
+					if ($('#rightCouplingControl:hover').length == 0) {
+						if (GUI.defaultZoomPanState('right', false, event)) return;
+					}
+				} else {
+					if ($('#leftCouplingControl:hover').length == 0) {
+						if (GUI.defaultZoomPanState('left', false, event)) return;
+					}
+				}
+			}
+
 			if (clickedObject) {
 				// Objects with restricted moving areas should get the "native" events
 				// Only if clicked on the moving area, e.g. actionbar the default event handling
@@ -484,13 +629,92 @@ GUI.initMouseHandler = function() {
 }
 
 
+
+
+/**
+ * set the final position of an arrow or line after creation via the selecting-points-method
+ */
+GUI.setFinalPosition = function(object){
+	
+	var title = $('#besideMouse').attr('title');
+	
+	var point = title.split(',');
+	
+	var x1 = parseInt(point[0])-15;
+	var y1 = parseInt(point[1])-35; 
+	var x2 = parseInt(point[2])-15;
+	var y2 = parseInt(point[3])-35; 
+		
+	var direction1;
+	var direction2;	
+	
+	var x;
+	var y;
+	var width;
+	var height;
+	
+	//calculate x,y, width and height
+	if(x1>x2){
+		x = x2;
+		width = x1-x2;
+		direction1 = false;
+	}
+	else{
+		x = x1;
+		width = x2-x1
+		direction1 = true;
+	}
+	if(y1>y2){
+		y = y2;
+		height = y1-y2;
+		direction2 = false;
+	}
+	else{
+		y = y1;
+		height = y2-y1;
+		direction2 = true;
+	}
+	
+	//calculate and set direction
+	if(direction1){
+		if(direction2){
+			object.setAttribute("direction", 1);
+		}
+		else{
+			object.setAttribute("direction", 4);
+		}
+	}
+	else{
+		if(direction2){
+			object.setAttribute("direction", 2);
+		}
+		else{
+			object.setAttribute("direction", 3);
+		}
+	}
+	
+	//set x,y, width and height
+	object.setAttribute('x', x);
+	object.setAttribute('y', y);
+	object.setAttribute('width', width);
+	object.setAttribute('height', height);
+	object.setViewWidth(width);
+	object.setViewHeight(height); 
+	
+	$('#besideMouse').attr('title', "");
+	
+}
+
+
+
+
 /**
  * get the topmost object at point x,y which is visible
  * @param {int} x x position
  * @param {int} y y position
  */
 GUI.getObjectAt = function(x,y) {
-	
+
 	var clickedObject = false;
 	
 	$.each(ObjectManager.getObjectsByLayer(), function(key, object) {
@@ -505,14 +729,10 @@ GUI.getObjectAt = function(x,y) {
 		}
 
 	});
-	
+
 	return clickedObject;
 	
 }
-
-
-
-
 
 
 /**
@@ -678,7 +898,13 @@ GUI.disconnected = function() {
 GUI.connected = function() {
 
 	if (GUI.relogin === true) {
+
+		if (GUI.couplingModeActive) {
+			GUI.closeCouplingMode();
+		}
+
 		GUI.relogin = false;
+
 		GUI.login();
 	}
 	
@@ -712,4 +938,22 @@ GUI.startNoAnimationTimer = function() {
  */
 GUI.confirm = function(message) {
 	return confirm(message);
+}
+
+/**
+ * set text which is displayed near to the cursor
+ */
+GUI.setCursorText = function(text){
+	$(document).mousemove(function(e){
+		var cpos = { top: e.pageY + 15, left: e.pageX + 15 };
+		$('#besideMouse').offset(cpos);
+		$('#besideMouse').html(text);	
+	});
+}
+
+/**
+ * get text which is displayed near to the cursor
+ */
+GUI.getCursorText = function(){
+	return $('#besideMouse').html();	
 }
